Back-End/hello-spring/src/main/java/com/example/demo/hellospring/service/MemberService.java
package com.example.demo.hellospring.service;

import java.util.List;
import java.util.Optional;

import com.example.demo.hellospring.domain.Member;
import com.example.demo.hellospring.repository.MemberRepository;

import org.springframework.beans.factory.annotation.Autowired;


public class MemberService {
    
    private final MemberRepository memberRepository;

    @Autowired
    public MemberService(MemberRepository memberRepository) {
        this.memberRepository = memberRepository;
    }

    /**
     * 회웡가입
     * @param member
     * @return
     */
    public Long join(Member member){
        //중복 회원 검증
        /* Optional<Member> result = memberRepository.findByName(member.getName());
        result.ifPresent(m->{
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        }); */
        validateDuplicateMember(member);
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        memberRepository.findByName(member.getName())
        .ifPresent(m->{
            throw new IllegalStateException("이미 존재하는 회원입니다.");
        });
    }

    /**
     * 전체 회원 조회
     */
    public List<Member> findMembers(){
        return memberRepository.findAll();
    }

    public Optional<Member> findOne(Long memberId){
        return memberRepository.findById(memberId);
    }
}
